+++ client/src/scripts/objects/Body.ts
@@ -1,4 +1,4 @@
-export default class Body extends Phaser.Physics.Arcade.Sprite {
+export default class BodySprite extends Phaser.Physics.Arcade.Sprite {
   constructor(scene: Phaser.Scene, x: number, y: number) {
     super(scene, x, y, "body.blue.1");
     scene.add.existing(this);