+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## 0.1.0
 
+- player look into walk direction
 - integrate matter js as physics engine
 - client can control player bodies
 - player synchronizes across clients